+++ src/IntegrationLibrary/Features/BloodBank/Service/IUserService.cs
@@ -13,6 +13,6 @@ namespace IntegrationLibrary.Features.BloodBank.Service
         IEnumerable<User> GetAll();
         User GetBy(string email);
         void ChangePassword(string email, ChangePasswordDTO dto);
-        User GetById(long id);
+        User GetById(int id);
     }
 }
