+++ src/IntegrationLibrary/Features/BloodBank/Service/UserService.cs
@@ -112,7 +112,7 @@ namespace IntegrationLibrary.Features.BloodBank.Service
             return _userRepository.GetBy(email);
         }
 
-        public User GetById(long id)
+        public User GetById(int id)
         {
             return _userRepository.GetById(id);
         }
